requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
click==7.1.2              # via flask
flask-cors==3.0.8         # via -r requirements.in
flask==1.1.2              # via -r requirements.in, flask-cors
itsdangerous==1.1.0       # via flask
jinja2==2.11.2            # via flask, sqla-raw
markupsafe==1.1.1         # via jinja2
marshmallow==3.6.1        # via -r requirements.in
psycopg2-binary==2.8.5    # via -r requirements.in
six==1.15.0               # via flask-cors
sqla-raw==0.1.1           # via -r requirements.in
sqlalchemy==1.3.17        # via sqla-raw
werkzeug==1.0.1           # via flask
